[서버] warning 처리 및 cannon fire 작업중

+++ Server/Single Thread ASIO/source/IngameData.h
@@ -6,14 +6,14 @@
 class cIngameData {
 public:
 	int					room_num_;
-	float				x_, y_, z_, r_;			// 캐릭터의 좌표
-	float				rx_, ry_, rz_;			// 캐릭터의 반경
+	double				x_, y_, z_, r_;			// 캐릭터의 좌표
+	double				rx_, ry_, rz_;			// 캐릭터의 반경
 	int					role_;					// 1~5 생존자, 6~7 살인마
 	int					hp_;					// 체력
 	int					before_hp_;				// 이전 체력 ( 데미지 검사용 )
 	std::string			user_name_;				// 유저 이름
 	bool				die_;					// 사망 여부
-	float				speed_;					// 이동 속도
+	double				speed_;					// 이동 속도
 	bool				jump_;					// 점프 여부
 	bool				interaction_;			// 상호작용 중인지 확인
 	int					fuse_;					// 지금 가지고있는 퓨즈의 인덱스, 없으면 -1
